chore: rename voting power route
