+++ packages/shared/routes/dashboard/governance/views/GovernanceDashboard.svelte
@@ -1,6 +1,6 @@
 <script lang="typescript">
     import { DashboardPane } from 'shared/components'
-    import { GovernanceVotingPower, GovernanceInfo, GovernanceEvents } from './'
+    import { GovernanceVotingWeight, GovernanceInfo, GovernanceEvents } from './'
     import { ParticipationEvent } from 'shared/lib/participation/types'
     import { WalletAccount } from 'shared/lib/typings/wallet'
 
@@ -11,7 +11,7 @@
 <div class="w-full h-full grid grid-cols-3 gap-x-4 min-h-0">
     <div class="h-full flex flex-col space-y-3 overflow-hidden">
         <DashboardPane classes="w-full flex-shrink-0 p-6">
-            <GovernanceVotingPower {account} />
+            <GovernanceVotingWeight {account} />
         </DashboardPane>
         <DashboardPane classes="w-full flex-grow p-6">
             <GovernanceInfo />
