+++ packages/shared/routes/dashboard/governance/views/GovernanceVotingWeight.svelte
